Implement `BoardBusToast`
